extract SupportIcon
